+++ src/main/java/com/umc/commonplant/domain/oauth/OAuthProvider.java
@@ -1,0 +1,5 @@
+package com.umc.commonplant.domain.oauth;
+
+public enum OAuthProvider {
+    KAKAO, APPLE
+}
