+++ src/main/java/com/umc/commonplant/domain/oauth/service/OAuthService.java
@@ -1,11 +1,9 @@
-package com.umc.commonplant.domain.user.service;
+package com.umc.commonplant.domain.oauth.service;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.umc.commonplant.domain.Jwt.JwtService;
-import com.umc.commonplant.domain.oauth.OAuthInfoResponse;
-import com.umc.commonplant.domain.user.dto.KakaoProfile;
-import com.umc.commonplant.domain.user.dto.UserDto;
+import com.umc.commonplant.domain.oauth.dto.KakaoProfile;
 import com.umc.commonplant.domain.user.entity.User;
 import com.umc.commonplant.domain.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
@@ -18,7 +16,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.util.MultiValueMap;
 import org.springframework.web.client.HttpClientErrorException;
 import org.springframework.web.client.RestTemplate;
-import org.springframework.web.multipart.MultipartFile;
 
 @Slf4j
 @Service
@@ -26,34 +23,22 @@ import org.springframework.web.multipart.MultipartFile;
 public class OAuthService {
     private final ObjectMapper objectMapper;
     private final UserRepository userRepository;
-    private final UserService userService;
     private final JwtService jwtService;
 
     public String oAuthLogin(String accessToken, String provider){
         String email = "";
-        String name = "";
-        String img = "";
-        MultipartFile mockImg = null;
-//        email = kakaoLogin(accessToken);
-        KakaoProfile kakaoProfile = kakaoLogin(accessToken);
-        email = kakaoProfile.getKakao_account().getEmail();
-        name = kakaoProfile.getProperties().getNickname();
-        img = kakaoProfile.getProperties().getProfile_image();
+        email = kakaoLogin(accessToken);
 
-        if(userRepository.countUserByEmail(email, provider) > 0){
+        if(userRepository.countUserByEmail(email, provider) > 0){ // 로그인 성공
             User user = userRepository.findByEmail(email, provider);
             String token = jwtService.createToken(user.getUuid());
             return token;
         }else{
-            //join
-            UserDto.join req = new UserDto.join();
-            req.setEmail(email);
-            req.setName(name);
-            req.setProvider("kakao");
-            return userService.joinUser(req, mockImg);
+            log.info("[REJECT]no info in DB");
+            throw new IllegalArgumentException(email);
         }
     }
-    public KakaoProfile kakaoLogin(String accessToken){
+    public String kakaoLogin(String accessToken){
         RestTemplate restTemplate = new RestTemplate();
         HttpHeaders headers = new HttpHeaders();
         HttpEntity<MultiValueMap<String, String>> request = new HttpEntity(headers);
@@ -72,6 +57,6 @@ public class OAuthService {
         }catch(JsonProcessingException e){
             log.info("[REJECT]kakaoMapper error");
         }
-        return kakaoProfile;
+        return kakaoProfile.getKakao_account().getEmail();
     }
 }
